refactor(tests): remove obsolete test_consolidate_results.py and clean up test_aggregator.py
- Deleted the test_consolidate_results.py file as it was no longer needed.
- Cleaned up test_aggregator.py by removing an unused test function related to ViewBuilder integration.
- Updated the remaining tests to ensure they function correctly without the removed components.
